limpieza de archivos y agregado de funciones
